Small fix for the getHistory/getHistoryIflx2 to decrease possible time difference for first and last items from the desired interval borders
Currently in this adapter in functions getHistory/getHistoryIflx2 the history data is requested from database  not only inside of the time interval, received by getHistory request, but, for "smooth" drawing, one value before and one after in the appropriate timeseries...
But, in some cases, the timestamp of this "additional" data-points can be significantly differ from begin and end values(borders) of requested time interval... Up to several days ...
To avoid it the time stamp of this data-points will be manually set as much close to the "borders" as possible. In proposed PR it is equal to 1 second.